+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Onboarding views
+
+### Changed
+ - Change default goals page to personal
 
 ## 0.0.2 - 2019-03-06